+++ SupportBank/Transaction.cs
@@ -12,9 +12,9 @@ namespace SupportBank
         public Person To { get; set; }
         public Person From { get; set; }
         public string Narrative { get; set; }
-        public float Amount { get; set; }
+        public decimal Amount { get; set; }
 
-        public Transaction(DateTime date, Person to, Person from, string narrative, float amount)
+        public Transaction(DateTime date, Person to, Person from, string narrative, decimal amount)
         {
             Date = date;
             To = to;
